requirements.txt
customtkinter
pymongo
opencv-python
face_recognition
pillow
numpy
reportlab

